fixing refs and moving test container class into test utils
