+++ apps/bfd-pipeline/bfd-pipeline-shared-test-utils/pom.xml
@@ -52,6 +52,11 @@
         <artifactId>slf4j-api</artifactId>
         <version>${slf4j.version}</version>
     </dependency>
+    <dependency>
+        <groupId>org.testcontainers</groupId>
+        <artifactId>testcontainers</artifactId>
+        <version>1.17.6</version>
+    </dependency>
 
 </dependencies>
