+++ apps/bfd-pipeline/bfd-pipeline-shared-test-utils/src/main/java/gov/cms/bfd/pipeline/MinioTestContainer.java
@@ -1,4 +1,4 @@
-package gov.cms.bfd.pipeline.sharedutils;
+package gov.cms.bfd.pipeline;
 
 import com.amazonaws.ClientConfiguration;
 import com.amazonaws.auth.AWSCredentials;
