+++ apps/bfd-pipeline/bfd-pipeline-shared-utils/pom.xml
@@ -114,11 +114,6 @@
 			<version>${mockito.version}</version>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
-			<groupId>org.testcontainers</groupId>
-			<artifactId>testcontainers</artifactId>
-			<version>1.17.6</version>
-		</dependency>
 	</dependencies>
 
 	<build>
